refactor: Moves regions to appropriate classes; confirms to naming convention.
