+++ lab/AboutDialog/AboutDialog/App.config
@@ -7,5 +7,5 @@
     gitHubAuthToken="7d42df1dd95aa9368b75fcb7883b11f428d27ca3"
     gitHubLogin="sharpenrocks"
     gitHubRepository="Sharpen"
-    vSMarketplaceId="ironcev.Sharpen" />
+    visualStudioMarketplaceId="ironcev.Sharpen" />
 </configuration>
