+++ lab/AboutDialog/AboutDialog/Extensions/BitmapExtensions.cs
@@ -15,12 +15,12 @@ namespace Sharpen
         /// <returns>System.Drawing.Bitmap</returns>
         public static  BitmapImage ConvertToBitmapImage(this Bitmap source)
         {
-            using (MemoryStream stream = new MemoryStream())
+            using (var stream = new MemoryStream())
             {
                 source.Save(stream, ImageFormat.Png);
 
                 stream.Position = 0;
-                BitmapImage result = new BitmapImage();
+                var result = new BitmapImage();
                 result.BeginInit();
                 // According to MSDN, "The default OnDemand cache option retains access to the stream until the image is needed."
                 // Force the bitmap to load right now so we can dispose the stream.
