+++ lab/AboutDialog/AboutDialog/Resources/Icon.VisualStudioMarketplace.png
